requirements.txt
apache-beam[gcp]>=2.16.0
tensorflow>=0.14.0,<2.0.0
scipy
google-resumable-media==0.5.0